Changaed all int to doubel on Api Models

+++ WeatherApiMvcApp/Models/ApiModels/Clouds.cs
@@ -7,6 +7,6 @@ namespace WeatherApiMvcApp.Models.ApiModels
 {
     public class Clouds
     {
-        public int all { get; set; }
+        public double all { get; set; }
     }
 }
