+++ WeatherApiMvcApp/Models/ApiModels/Main.cs
@@ -8,8 +8,8 @@ namespace WeatherApiMvcApp.Models.ApiModels
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
-        public int humidity { get; set; }
+        public double pressure { get; set; }
+        public double humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
     }
