+++ WeatherApiMvcApp/Models/ApiModels/RootObject.cs
@@ -11,13 +11,13 @@ namespace WeatherApiMvcApp.Models.ApiModels
         public List<Weather> weather { get; set; }
         public string @base { get; set; }
         public Main main { get; set; }
-        public int visibility { get; set; }
+        public double visibility { get; set; }
         public Wind wind { get; set; }
         public Clouds clouds { get; set; }
-        public int dt { get; set; }
+        public double dt { get; set; }
         public Sys sys { get; set; }
-        public int id { get; set; }
+        public double id { get; set; }
         public string name { get; set; }
-        public int cod { get; set; }
+        public double cod { get; set; }
     }
 }
