+++ WeatherApiMvcApp/Models/ApiModels/Wind.cs
@@ -8,6 +8,6 @@ namespace WeatherApiMvcApp.Models.ApiModels
     public class Wind
     {
         public double speed { get; set; }
-        public int deg { get; set; }
+        public double deg { get; set; }
     }
 }
